processor/requirements.txt
py4j==0.10.9.7
pyspark==3.5.0
python-dotenv==1.0.0
